requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
Deprecated==1.2.12
idna==3.2
mutagen==1.45.1
numpy==1.21.2
pandas==1.3.2
pycparser==2.20
pycryptodomex==3.12.0
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
python-dateutil==2.8.2
pytz==2021.1
pyupbit==0.2.19
requests==2.26.0
six==1.16.0
slack-sdk==3.15.2
urllib3==1.26.6
websockets==9.1
wrapt==1.12.1
requests==2.26.0
